apiextensions: move conversion tests into own package
