+++ staging/src/k8s.io/apiextensions-apiserver/test/integration/conversion/webhook.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package convert
+package conversion
 
 import (
 	"crypto/tls"
